requirements.txt
flask
flask-sqlalchemy
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
google-generativeai==0.3.1
SQLAlchemy==1.4.31
pytest==7.1.2
gunicorn==20.1.0
Werkzeug==2.0.2
Jinja2==3.0.3
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.3